Name stack resources uniquely with a uuid so conflicts don't arise when users deploy this project multiple times to the same region.
Details: In the base stack, I use a custom resource to create a random string for the uuid. I then pass that as ShortUUID to all the nested stacks. This way we can ensure the video and image workflows are named with a uuid suffix, and that uuid is used in the front-end when executing workflows.