Derive the s3 bucket name for web access logging with a custom resource instead of a cloud formation macro.
Details: Cloud formation macros cannot be used if we want to allow a stack to be deployed multiple times in the same region. This is because the Transform functions that call macros must reference the macro by a name which is resolved during compile time. You are not allowed to use any variables in the macro name property for the Transform.